refactor: coach, crew 사용하는 곳 리팩토링
