+++ backend/src/main/java/com/woowacourse/ternoko/domain/AvailableDateTime.java
@@ -8,6 +8,9 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
+
+import com.woowacourse.ternoko.domain.member.Coach;
+
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
