--- backend/src/main/java/com/woowacourse/ternoko/domain/Crew.java
@@ -1,12 +1,0 @@
-package com.woowacourse.ternoko.domain;
-
-import javax.persistence.DiscriminatorValue;
-import javax.persistence.Entity;
-
-import lombok.NoArgsConstructor;
-
-@Entity
-@NoArgsConstructor
-@DiscriminatorValue("CREW")
-public class Crew extends Member{
-}
