+++ backend/src/main/java/com/woowacourse/ternoko/repository/CoachRepository.java
@@ -2,7 +2,7 @@ package com.woowacourse.ternoko.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.woowacourse.ternoko.domain.Coach;
+import com.woowacourse.ternoko.domain.member.Coach;
 
 public interface CoachRepository extends JpaRepository<Coach, Long> {
 }
