+++ backend/src/main/java/com/woowacourse/ternoko/repository/CrewRepository.java
@@ -2,7 +2,7 @@ package com.woowacourse.ternoko.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.woowacourse.ternoko.domain.Crew;
+import com.woowacourse.ternoko.domain.member.Crew;
 
 public interface CrewRepository extends JpaRepository<Crew, Long> {
 }
